requirements.txt
flask
transformers
torch==2.0.0
qdrant-client
pandas
python-dotenv
numpy==1.24.4
venv/
